+++ crates/runner-nordic/src/storage.rs
@@ -42,8 +42,8 @@ macro_rules! take_storage {
             static mut $start: u32;
             static mut $end: u32;
         }
-        let start = unsafe { addr_of_mut!($start) as *mut u8 };
-        let end = unsafe { addr_of_mut!($end) as usize };
+        let start = addr_of_mut!($start) as *mut u8;
+        let end = addr_of_mut!($end) as usize;
         let length = end.checked_sub(start as usize).unwrap();
         assert_eq!(length % PAGE_SIZE, 0);
         unsafe { slice::from_raw_parts_mut(start, length) }