+++ crates/runner-host/src/board/debug.rs
@@ -37,6 +37,9 @@ impl board::debug::Api for Impl {
     }
 
     fn exit(success: bool) -> ! {
+        for cleanup in std::mem::take(&mut *crate::CLEANUP.lock().unwrap()) {
+            cleanup();
+        }
         std::process::exit(if success { 0 } else { 1 })
     }
 }