+++ crates/cli-tools/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Change the behavior of `fs::copy_if_changed()` to keep an original source
 
+### Patch
+
+- Update dependencies
+
 ## 0.1.0
 
 <!-- Increment to skip CHANGELOG.md test: 0 -->