+++ crates/cli-tools/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## 0.1.1-git
+## 0.2.0-git
+
+### Major
+
+- Change all actions requiring a connection to use async
+- Change the return type of `ConnectionOptions::connect()`
+- Change `action::PlatformList::run()` and remove `action::PlatformInfo`
 
 ### Minor